COMMON: - Add the Aggregating iterators with the `next(Aggregator)` method and split   it out from the TypedTimeSeriesIterator to make it cleaner and useful when   only some nodes support it.
CORE:
- Implement the above.
