+++ common/src/main/java/net/opentsdb/data/AggregatingTypedTimeSeriesIterator.java
@@ -1,5 +1,5 @@
 // This file is part of OpenTSDB.
-// Copyright (C) 2018-2020  The OpenTSDB Authors.
+// Copyright (C) 2020  The OpenTSDB Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,23 +14,21 @@
 // limitations under the License.
 package net.opentsdb.data;
 
-import com.google.common.reflect.TypeToken;
-
-import java.io.Closeable;
-import java.util.Iterator;
-
 /**
- * An iterator for {@link TimeSeriesValue}s that lets us determine the
- * type of the underlying data without having to resolve the types.
+ * An interface that extends the iterator with a {@link #next(Aggregator)} method
+ * that will populate the given aggregator with the values at the proper 
+ * index. This can only be used for downsampled, normalized data.
  * 
  * @since 3.0
  */
-public interface TypedTimeSeriesIterator<T extends TimeSeriesDataType> 
-    extends Iterator<TimeSeriesValue<T>>, Closeable {
+public interface AggregatingTypedTimeSeriesIterator<T extends TimeSeriesDataType> 
+    extends TypedTimeSeriesIterator<T> {
 
   /**
-   * @return The non-null type of data returned in the iterator.
+   * Runs through the data and adds the results to the aggregator at the proper
+   * index.
+   * @param aggregator A non-null aggregator to populate.
    */
-  public TypeToken<T> getType();
+  public void next(final Aggregator aggregator);
   
 }
