+++ common/src/main/java/net/opentsdb/query/AggregatingQueryIterator.java
@@ -1,5 +1,5 @@
 // This file is part of OpenTSDB.
-// Copyright (C) 2018-2020  The OpenTSDB Authors.
+// Copyright (C) 2020  The OpenTSDB Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,25 +12,16 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package net.opentsdb.data;
+package net.opentsdb.query;
 
-import com.google.common.reflect.TypeToken;
-
-import java.io.Closeable;
-import java.util.Iterator;
+import net.opentsdb.data.AggregatingTypedTimeSeriesIterator;
 
 /**
- * An iterator for {@link TimeSeriesValue}s that lets us determine the
- * type of the underlying data without having to resolve the types.
+ * An iterator generated at query time over time series values that can popluate
+ * an aggregator on calls to {@link #next(net.opentsdb.data.Aggregator)}.
  * 
  * @since 3.0
  */
-public interface TypedTimeSeriesIterator<T extends TimeSeriesDataType> 
-    extends Iterator<TimeSeriesValue<T>>, Closeable {
+public interface AggregatingQueryIterator extends AggregatingTypedTimeSeriesIterator {
 
-  /**
-   * @return The non-null type of data returned in the iterator.
-   */
-  public TypeToken<T> getType();
-  
 }
