+++ common/src/main/java/net/opentsdb/query/QueryIterator.java
@@ -14,7 +14,6 @@
 // limitations under the License.
 package net.opentsdb.query;
 
-import net.opentsdb.data.TimeSeriesDataType;
 import net.opentsdb.data.TypedTimeSeriesIterator;
 
 /**
